+++ truffle/src/com.oracle.truffle.api.test/src/com/oracle/truffle/api/test/polyglot/ProcessEnvironmentTest.java
@@ -40,7 +40,6 @@
  */
 package com.oracle.truffle.api.test.polyglot;
 
-import java.util.HashMap;
 import java.util.Map;
 import org.graalvm.polyglot.Context;
 import org.graalvm.polyglot.EnvironmentAccess;
@@ -51,19 +50,10 @@ public class ProcessEnvironmentTest extends AbstractPolyglotTest {
 
     @Test
     public void testEnvironmentAccessNone() {
-        setupEnv(Context.newBuilder().allowEnvironmentAccess(EnvironmentAccess.NONE).environment("k1", "v1").environment("k2", "v2").build());
-        Assert.assertNull(languageEnv.getEnvironment("k1"));
-        Assert.assertNull(languageEnv.getEnvironment("k2"));
+        setupEnv(Context.newBuilder().allowEnvironmentAccess(EnvironmentAccess.NONE).build());
         Map<String, String> env = languageEnv.getEnvironment();
-        Assert.assertNotNull(env);
         Assert.assertTrue(env.isEmpty());
         try {
-            languageEnv.setEnvironment("k3", "v3");
-            Assert.fail("Expected SecurityException.");
-        } catch (SecurityException se) {
-            // expected
-        }
-        try {
             env.clear();
             Assert.fail("Environment map must be unmodifiable.");
         } catch (UnsupportedOperationException e) {
@@ -72,46 +62,10 @@ public class ProcessEnvironmentTest extends AbstractPolyglotTest {
     }
 
     @Test
-    public void testEnvironmentAccessRead() {
-        setupEnv(Context.newBuilder().allowEnvironmentAccess(EnvironmentAccess.READ).environment("k1", "v1").environment("k2", "v2").build());
-        Assert.assertEquals("v1", languageEnv.getEnvironment("k1"));
-        Assert.assertEquals("v2", languageEnv.getEnvironment("k2"));
+    public void testEnvironmentInherit() {
+        setupEnv(Context.newBuilder().allowEnvironmentAccess(EnvironmentAccess.INHERIT).build());
         Map<String, String> env = languageEnv.getEnvironment();
-        Map<String, String> expected = new HashMap<>();
-        expected.put("k1", "v1");
-        expected.put("k2", "v2");
-        Assert.assertEquals(expected, env);
-        try {
-            languageEnv.setEnvironment("k3", "v3");
-            Assert.fail("Expected SecurityException.");
-        } catch (SecurityException se) {
-            // expected
-        }
-        try {
-            env.clear();
-            Assert.fail("Environment map must be unmodifiable.");
-        } catch (UnsupportedOperationException e) {
-            // expected
-        }
-    }
-
-    @Test
-    public void testEnvironmentAccessReadWrite() {
-        setupEnv(Context.newBuilder().allowEnvironmentAccess(EnvironmentAccess.ALL).environment("k1", "v1").environment("k2", "v2").build());
-        Assert.assertEquals("v1", languageEnv.getEnvironment("k1"));
-        Assert.assertEquals("v2", languageEnv.getEnvironment("k2"));
-        Map<String, String> env = languageEnv.getEnvironment();
-        Map<String, String> expected = new HashMap<>();
-        expected.put("k1", "v1");
-        expected.put("k2", "v2");
-        Assert.assertEquals(expected, env);
-        languageEnv.setEnvironment("k3", "v3");
-        Assert.assertEquals("v3", languageEnv.getEnvironment("k3"));
-        env = languageEnv.getEnvironment();
-        expected = new HashMap<>();
-        expected.put("k1", "v1");
-        expected.put("k2", "v2");
-        expected.put("k3", "v3");
+        Map<String, String> expected = System.getenv();
         Assert.assertEquals(expected, env);
         try {
             env.clear();